Enabled SSR for previews (#10)
The issue with social media previews seems to be that the <meta /> tags aren't being loaded into the DOM because the page itself is being rendered at the client side. In other words, as the browser requests for a page, it will get some JS as a response which it renders by itself. As such, only after rendering would the Title component in src/components/Head.jsx be added into the DOM. This in turn, won't allow any other site to read the <meta /> tags.

In order to fix this issue, this commit adds server side rendering for the Title component alone so that the HTML would be present in the DOM when someone requests for the page.